pyproject.toml
[build-system]
requires = [
  "setuptools >= 70.0",
  "wheel"                  # to build wheels
]
build-backend = "setuptools.build_meta"

[project]
name = "skytraffic"
version = "0.5.0"
description = "Your package description here"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Weijiang Xiong", email = "weijiangxiong1998@gmail.com" },
]
dependencies = [
  "torch",
  "torch_geometric",
  "torchvision",
  "matplotlib",
  "numpy",
  "seaborn",
  "einops",
  "omegaconf",
  "tabulate",
  "termcolor",
  "tqdm",
  "scipy",
  "hydra-core",
  "tensorboard",
]

[project.optional-dependencies]
vis = [
  "geopandas",
  "contextily",
]

[tool.setuptools]
# If you have one top-level package
packages = ["skytraffic"]

[tool.ruff]
target-version = "py311"